General Linear Model

General Linear Model. Estimates models using `lm()` function and provides options to facilitate estimation of interactions, simple slopes, simple effects, post-hoc tests, contrast analysis, effect size indexes and visualization of the results.

gamljGlm(
  data,
  formula,
  dep = NULL,
  factors = NULL,
  covs = NULL,
  modelTerms = NULL,
  fixedIntercept = TRUE,
  showParamsCI = TRUE,
  paramCIWidth = 95,
  contrasts = NULL,
  showRealNames = TRUE,
  showContrastCode = FALSE,
  plotHAxis = NULL,
  plotSepLines = NULL,
  plotSepPlots = NULL,
  plotRaw = FALSE,
  plotDvScale = FALSE,
  plotError = "none",
  ciWidth = 95,
  postHoc = NULL,
  eDesc = FALSE,
  eCovs = FALSE,
  simpleVariable = NULL,
  simpleModerator = NULL,
  simple3way = NULL,
  simpleScale = "mean_sd",
  cvalue = 1,
  percvalue = 25,
  simpleScaleLabels = "labels",
  postHocCorr = list("bonf"),
  scaling = NULL,
  effectSize = list("beta", "etap"),
  homoTest = FALSE,
  qq = FALSE,
  normTest = FALSE,
  normPlot = FALSE,
  residPlot = FALSE,
  interceptInfo = FALSE,
  effectSizeInfo = FALSE,
  dep_scale = "none"
)

Arguments

data

the data as a data frame

formula

(optional) an object of class formula (or one that can be coerced to that class): a symbolic description of the model to be fitted.

dep

a string naming the dependent variable; the variable must be numeric. Not needed if formula is used.

factors

a vector of strings naming the fixed factors. Not needed if formula is used.

covs

a vector of strings naming the covariates (continuous variables). Not needed if formula is used.

modelTerms

a list of character vectors describing fixed effects terms. Not needed if formula is used.

fixedIntercept

TRUE (default) or FALSE, estimates fixed intercept. Not needed if formula is used.

showParamsCI

TRUE (default) or FALSE , parameters CI in tables

paramCIWidth

a number between 50 and 99.9 (default: 95) specifying the confidence interval width for the parameter estimates

contrasts

a named vector of the form c(var1='type', var2='type2') specifying the type of contrast to use, one of 'deviation', 'simple', 'dummy', 'difference', 'helmert', 'repeated' or 'polynomial'. If NULL, 'simple' is used. Can also be passed as a list of list of the form list(list(var='var1',type='type1')).

showRealNames

TRUE or FALSE (default), shows raw names of the contrasts variables

showContrastCode

TRUE or FALSE (default), shows contrast coefficients tables

plotHAxis

a string naming the variable placed on the horizontal axis of the plot

plotSepLines

a string naming the variable represented as separate lines in the plot

plotSepPlots

a string naming the variable defining the levels for multiple plots

plotRaw

TRUE or FALSE (default), plot raw data along the predicted values

plotDvScale

TRUE or FALSE (default), set the Y-axis range equal to the range of the observed values.

plotError

'none' (default), 'ci', or 'se'. Use no error bars, use confidence intervals, or use standard errors on the plots, respectively.

ciWidth

a number between 50 and 99.9 (default: 95) specifying the confidence interval width for the plots.

postHoc

a rhs formula with the terms specifying the table to apply the comparisons (of the form '~x+x:z'

eDesc

TRUE or FALSE (default), provide lsmeans statistics

eCovs

TRUE or FALSE (default), provide lsmeans statistics conditioned to different values of the continuous variables in the model. Which levels of the continuous variable should be used is set by the simpleScale option.

simpleVariable

The variable for which the simple effects (slopes) are computed

simpleModerator

the variable that provides the levels at which the simple effects are computed

simple3way

a moderator of the two-way interaction which is probed

simpleScale

'mean_sd' (default), 'custom' , or 'percent'. Use to condition the covariates (if any)

cvalue

how many st.deviations around the means used to condition simple effects and plots. Used if simpleScale='mean_sd'

percvalue

offsett (number of percentiles) around the median used to condition simple effects and plots. Used if simpleScale='percent'

simpleScaleLabels

how the levels of a continuous moderator should appear in tables and plots: labels, values and values_labels.

postHocCorr

one or more of 'none', 'bonf', or 'holm'; provide no, Bonferroni, and Holm Post Hoc corrections respectively.

scaling

a named vector of the form c(var1='type1',var2='type2'). Types are 'centered' to the mean, 'standardized', log-transformed 'log' or 'none'. 'none' leaves the variable as it is. It can also be passed as a list of lists.

effectSize

a list of effect sizes to print out. They can be: 'eta' for eta-squared, 'etap' for partial eta-squared, 'omega' for omega-squared, 'omegap' for partial omega-squared 'epsilon' for epsilon-squared 'epsilonp' for partial epsilon-squared, and 'beta' for standardized coefficients (betas). Default is beta and etap.

homoTest

TRUE or FALSE (default), perform homogeneity tests

qq

TRUE or FALSE (default), provide a Q-Q plot of residuals

normTest

TRUE or FALSE (default), provide a test for normality of residuals

normPlot

TRUE or FALSE (default), provide a histogram of residuals superimposed by a normal distribution

residPlot

TRUE or FALSE (default), provide a scatterplot of the residuals against predicted

interceptInfo

TRUE or FALSE (default), provide ìnformation about the intercept (F test, effect size indexes)

effectSizeInfo

TRUE or FALSE (default), provide ìnformation about the effect size indexes

dep_scale

Re-scale the dependent variable. It can be 'centered' to the mean, 'standardized', log-transformed 'log' or 'none'. 'none' (default) leaves the variable as it is.

Value

A results object of class gamlj, gamlj,gamljGlmResults containing:

results$model					The underlying lm object

results$info					a table

results$main$interceptTable					a table of information for the model intercept

results$main$anova					a table of ANOVA results

results$main$effectSizeTable					a table of effect size indeces

results$main$fixed					a table

results$main$contrastCodeTables					an array of contrast coefficients tables

results$postHocs					an array of post-hoc tables

results$simpleEffects$Anova					a table of ANOVA for simple effects

results$simpleEffects$Params					a table

results$emeansTables					an array of predicted means tables

results$descPlot					a descriptives plot

results$descPlots					an array of results plots

results$assumptions$homoTest					a table of homogeneity tests

results$assumptions$normTest					a table of normality tests

results$assumptions$qq					a q-q plot

results$assumptions$normPlot					Residual histogram

results$assumptions$residPlot					Residual Predicted plot

Tables can be converted to data frames with asDF or as.data.frame. For example:

results$info$asDF

as.data.frame(results$info)

Examples

data('ToothGrowth')
gamlj::gamljGlm(formula = len ~ supp,  data = ToothGrowth)
#> Registered S3 method overwritten by 'gamlj':
#>   method     from  
#>   plot.gamlj GAMLj3
#> 
#>  GENERAL LINEAR MODEL
#> 
#>  Model Info                                    
#>  ───────────────────────────────────────────── 
#>    Info                                        
#>  ───────────────────────────────────────────── 
#>    Estimate          Linear model fit by OLS   
#>    Call              len ~ 1 + supp            
#>    R-squared                      0.05948365   
#>    Adj. R-squared                 0.04326785   
#>  ───────────────────────────────────────────── 
#> 
#> 
#>  MODEL RESULTS
#> 
#>  ANOVA Omnibus tests                                                    
#>  ────────────────────────────────────────────────────────────────────── 
#>                 SS           df    F           p            η²p         
#>  ────────────────────────────────────────────────────────────────────── 
#>    Model         205.3500     1    3.668253    0.0603934    0.0594836   
#>    supp          205.3500     1    3.668253    0.0603934    0.0594836   
#>    Residuals    3246.8593    58                                         
#>    Total        3452.2093    59                                         
#>  ────────────────────────────────────────────────────────────────────── 
#> 
#> 
#>  Fixed Effects Parameter Estimates                                                                                                  
#>  ────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────── 
#>    Names          Effect         Estimate     SE           Lower        Upper         β             df    t            p            
#>  ────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────── 
#>    (Intercept)    (Intercept)    18.813333    0.9659221    16.879830    20.7468365     0.0000000    58    19.477070    < .0000001   
#>    supp1          VC - OJ        -3.700000    1.9318443    -7.567006     0.1670064    -0.4837034    58    -1.915268     0.0603934   
#>  ────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────── 
#>