Generalized Linear Models

Generalized Linear Models. Estimates models using `glm()` function and similar functions. It provides options to facilitate estimation of interactions, simple slopes, simple effects, post-hoc tests, contrast analysis, effect size indexes and visualization of the results.

gamljGzlm(
  data,
  dep,
  factors = NULL,
  covs = NULL,
  modelTerms = NULL,
  fixedIntercept = TRUE,
  showParamsCI = FALSE,
  showExpbCI = TRUE,
  paramCIWidth = 95,
  contrasts = NULL,
  showRealNames = TRUE,
  showContrastCode = FALSE,
  plotHAxis = NULL,
  plotSepLines = NULL,
  plotSepPlots = NULL,
  plotRaw = FALSE,
  plotDvScale = FALSE,
  plotError = "none",
  ciWidth = 95,
  postHoc = NULL,
  eDesc = FALSE,
  eCovs = FALSE,
  simpleVariable = NULL,
  simpleModerator = NULL,
  simple3way = NULL,
  simpleScale = "mean_sd",
  cvalue = 1,
  percvalue = 25,
  simpleScaleLabels = "labels",
  postHocCorr = list("bonf"),
  scaling = NULL,
  effectSize = list("expb"),
  modelSelection = "linear",
  custom_family = "gaussian",
  custom_link = "identity",
  formula
)

Arguments

data

the data as a data frame

dep

a string naming the dependent variable from data, variable can be numeric or factor

factors

a vector of strings naming the fixed factors from data

covs

a vector of strings naming the covariates from data

modelTerms

a list of character vectors describing fixed effects terms

fixedIntercept

TRUE (default) or FALSE, estimates fixed intercept

showParamsCI

TRUE or FALSE (default), parameters CI in table

showExpbCI

TRUE (default) or FALSE , exp(B) CI in table

paramCIWidth

a number between 50 and 99.9 (default: 95) specifying the confidence interval width for the parameter estimates

contrasts

a named vector of the form c(var1='type', var2='type2') specifying the type of contrast to use, one of 'deviation', 'simple', 'dummy', 'difference', 'helmert', 'repeated' or 'polynomial'. If NULL, 'simple' is used. Can also be passed as a list of list of the form list(list(var='var1',type='type1')).

showRealNames

TRUE or FALSE (default), provide raw names of the contrasts variables

showContrastCode

TRUE or FALSE (default), provide contrast coefficients tables

plotHAxis

a string naming the variable placed on the horizontal axis of the plot

plotSepLines

a string naming the variable represented as separate lines on the plot

plotSepPlots

a string naming the variable to separate over to form multiple plots

plotRaw

TRUE or FALSE (default), provide descriptive statistics

plotDvScale

.

plotError

'none', 'ci' (default), or 'se'. Use no error bars, use confidence intervals, or use standard errors on the plots, respectively

ciWidth

a number between 50 and 99.9 (default: 95) specifying the confidence interval width

postHoc

a list of terms to perform post-hoc tests on

eDesc

TRUE or FALSE (default), provide lsmeans statistics

eCovs

TRUE or FALSE (default), provide lsmeans statistics

simpleVariable

The variable for which the simple effects (slopes) are computed

simpleModerator

the variable that provides the levels at which the simple effects computed

simple3way

a moderator of the two-way interaction which is probed

simpleScale

'mean_sd' (default), 'custom' , or 'custom_percent'. Use to condition the covariates (if any)

cvalue

offset value for conditioning

percvalue

offset value for conditioning

simpleScaleLabels

style for presenting condition values of a moderator. It can be 'labels' (default), 'values' or 'labels_values' for both.

postHocCorr

one or more of 'none', 'bonf', or 'holm'; provide no, Bonferroni, and Holm Post Hoc corrections respectively

scaling

a named vector of the form c(var1='type1',var2='type2'). Types are 'centered' to the mean, 'standardized', log-transformed 'log' or 'none'. 'none' leaves the variable as it is. It can also be passed as a list of lists.

effectSize

.

modelSelection

Select the generalized linear model: linear,poisson,logistic,multinomial

custom_family

Distribution family for the custom model, accepts gaussian, binomial, gamma and inverse_gaussian .

custom_link

Distribution family for the custom model, accepts identity, log and inverse, onemu2 (for 1/mu^2).

formula

(optional) the formula to use, see the examples

Value

A results object containing:

results$model					The underlying glm object

results$info					a table

results$main$anova					a table of ANOVA results

results$main$fixed					a table

results$main$relativerisk					a table

results$main$contrastCodeTables					an array of contrast coefficients tables

results$postHocs					an array of post-hoc tables

results$simpleEffects$Anova					a table of ANOVA for simple effects

results$simpleEffects$Params					a table

results$emeansTables					an array of predicted means tables

results$descPlot					a descriptives plot

results$descPlots					an array of results plots

Tables can be converted to data frames with asDF or as.data.frame. For example:

results$info$asDF

as.data.frame(results$info)

Examples

data<-emmeans::neuralgia
 gamlj::gamljGzlm(
           formula = Pain ~ Duration,
           data = data,
            modelSelection = "logistic")
#> Waiting for profiling to be done...
#> 
#>  GENERALIZED LINEAR MODEL
#> 
#>  Model Info                                                                     
#>  ────────────────────────────────────────────────────────────────────────────── 
#>    Info              Value            Comment                                   
#>  ────────────────────────────────────────────────────────────────────────────── 
#>    Model Type        Logistic         Model for binary y                        
#>    Call              glm              Pain ~ 1 + Duration                       
#>    Link function     Logit            Log of the odd of y=1 over y=0            
#>    Direction         P(y=1)/P(y=0)    P( Pain = Yes ) / P( Pain = No )          
#>    Distribution      Binomial         Dichotomous event distribution of y       
#>    R-squared            0.01983700    Proportion of reduction of error          
#>    AIC                 83.88600000    Less is better                            
#>    BIC                 88.07500000    Less is better                            
#>    Deviance            79.88641275    Less is better                            
#>    Residual DF         58.00000000                                              
#>    Chi-squared/DF       1.02844428    Overdispersion indicator                  
#>    Converged         yes              Whether the estimation found a solution   
#>  ────────────────────────────────────────────────────────────────────────────── 
#> 
#> 
#>  MODEL RESULTS
#> 
#>  Loglikelihood ratio tests                   
#>  ─────────────────────────────────────────── 
#>                X²          df    p           
#>  ─────────────────────────────────────────── 
#>    Duration    1.616779     1    0.2035415   
#>  ─────────────────────────────────────────── 
#> 
#> 
#>  Parameter Estimates                                                                                          
#>  ──────────────────────────────────────────────────────────────────────────────────────────────────────────── 
#>    Names          Estimate       SE            exp(B)       Lower        Upper       z            p           
#>  ──────────────────────────────────────────────────────────────────────────────────────────────────────────── 
#>    (Intercept)    -0.34728878    0.26600165    0.7066012    0.4145088    1.183753    -1.305589    0.1916924   
#>    Duration       -0.02971640    0.02390160    0.9707208    0.9239678    1.015884    -1.243281    0.2137643   
#>  ──────────────────────────────────────────────────────────────────────────────────────────────────────────── 
#>