Mixed Model

Mixed Linear Model. Estimates models using `me4::lmer()` function and provides options to facilitate estimation of interactions, simple slopes, simple effects, post-hoc tests, contrast analysis, effect size indexes and visualization of the results.

gamljMixed(
  data,
  dep = NULL,
  factors = NULL,
  covs = NULL,
  modelTerms = NULL,
  fixedIntercept = TRUE,
  showParamsCI = TRUE,
  paramCIWidth = 95,
  ciRE = FALSE,
  contrasts = NULL,
  showRealNames = TRUE,
  showContrastCode = FALSE,
  plotHAxis = NULL,
  plotSepLines = NULL,
  plotSepPlots = NULL,
  plotRaw = FALSE,
  plotDvScale = FALSE,
  plotError = "none",
  ciWidth = 95,
  postHoc = NULL,
  eDesc = FALSE,
  eCovs = FALSE,
  simpleVariable = NULL,
  simpleModerator = NULL,
  simple3way = NULL,
  simpleScale = "mean_sd",
  cvalue = 1,
  percvalue = 25,
  simpleScaleLabels = "labels",
  postHocCorr = list("bonf"),
  scaling = NULL,
  dep_scale = "none",
  cluster = NULL,
  randomTerms = list(list()),
  correlatedEffects = "corr",
  reml = TRUE,
  lrtRandomEffects = FALSE,
  plotRandomEffects = FALSE,
  cimethod = "wald",
  dfmethod = "Satterthwaite",
  qq = FALSE,
  normTest = FALSE,
  normPlot = FALSE,
  residPlot = FALSE,
  clusterBoxplot = FALSE,
  randHist = FALSE,
  formula
)

Arguments

data

the data as a data frame

dep

a string naming the dependent variable from data, variable must be numeric

factors

a vector of strings naming the fixed factors from data

covs

a vector of strings naming the covariates from data

modelTerms

a list of character vectors describing fixed effects terms

fixedIntercept

TRUE (default) or FALSE, estimates fixed intercept

showParamsCI

TRUE (default) or FALSE , parameters CI in table

paramCIWidth

a number between 50 and 99.9 (default: 95) specifying the confidence interval width for the parameter estimates

ciRE

TRUE or FALSE (default), produce random effects confidence intervals based on profile method

contrasts

a named vector of the form c(var1='type', var2='type2') specifying the type of contrast to use, one of 'deviation', 'simple', 'dummy', 'difference', 'helmert', 'repeated' or 'polynomial'. If NULL, 'simple' is used. Can also be passed as a list of list of the form list(list(var='var1',type='type1')).

showRealNames

TRUE or FALSE (default), provide raw names of the contrasts variables

showContrastCode

TRUE or FALSE (default), provide contrast coefficients tables

plotHAxis

a string naming the variable placed on the horizontal axis of the plot

plotSepLines

a string naming the variable represented as separate lines on the plot

plotSepPlots

the variable for whose levels multiple plots are computed

plotRaw

TRUE or FALSE (default), provide descriptive statistics

plotDvScale

TRUE or FALSE (default), scale the plot Y-Axis to the max and the min of the dependent variable observed scores.

plotError

'none', 'ci' (default), or 'se'. Use no error bars, use confidence intervals, or use standard errors on the plots, respectively

ciWidth

a number between 50 and 99.9 (default: 95) specifying the confidence interval width

postHoc

a list of terms to perform post-hoc tests on

eDesc

TRUE or FALSE (default), provide lsmeans statistics

eCovs

TRUE or FALSE (default), provide lsmeans statistics

simpleVariable

The variable for which the simple effects (slopes) are computed

simpleModerator

the variable that provides the levels at which the simple effects computed

simple3way

a moderator of the two-way interaction which is probed

simpleScale

'mean_sd' (default), 'custom' , or 'custom_percent'. Use to condition the covariates (if any)

cvalue

offset value for conditioning

percvalue

offset value for conditioning

simpleScaleLabels

decide the labeling of simple effects in tables and plots. labels indicates that only labels are used, such as Mean and Mean + 1 SD. values uses the actual values as labels. values_labels uses both.

postHocCorr

one or more of 'none', 'bonf', or 'holm'; provide no, Bonferroni, and Holm Post Hoc corrections respectively

scaling

a named vector of the form c(var1='type1',var2='type2'). Types are 'centered' to the mean, 'clusterbasedcentered' to the mean of each cluster, 'standardized', 'clusterbasedstandardized' standardized within each cluster, log-transformed 'log', or 'none'. 'none' leaves the variable as it is. It can also be passed as a list of lists.

dep_scale

Re-scale the dependent variable. It could be 'centered' to the mean, 'clusterbasedcentered' to the mean of each cluster, 'standardized', 'clusterbasedstandardized' standardized within each cluster, log-transformed 'log', or 'none' (default).

cluster

a vector of strings naming the clustering variables from data

randomTerms

a list of lists specifying the models random effects.

correlatedEffects

'nocorr', 'corr' (default), or 'block'. When random effects are passed as list of length 1, it decides whether the effects should be correlated, non correlated. If 'randomTerms' is a list of lists of length > 1, the option is automatially set to 'block'. The option is ignored if the model is passed using formula.

reml

TRUE (default) or FALSE, should the Restricted ML be used rather than ML

lrtRandomEffects

TRUE or FALSE (default), LRT for the random effects

plotRandomEffects

TRUE or FALSE (default), add random effects predicted values in the plot

cimethod

.

dfmethod

.

qq

TRUE or FALSE (default), provide a Q-Q plot of residuals

normTest

TRUE or FALSE (default), provide a test for normality of residuals

normPlot

TRUE or FALSE (default), provide a histogram of residuals superimposed by a normal distribution

residPlot

TRUE or FALSE (default), provide a scatterplot of the residuals against predicted

clusterBoxplot

TRUE or FALSE (default), provide a boxplot of random effects by the first defined clustering variable

randHist

TRUE or FALSE (default), provide histogram of random Coefficients

formula

(optional) the formula to use, see the examples#'

Value

A results object containing:

results$model					The underlying lm object

results$info					a table

results$main$anova					a table of ANOVA results

results$main$fixed					a table

results$main$random					a table

results$main$randomCov					a table

results$main$lrtRandomEffectsTable					a table

results$main$contrastCodeTables					an array of contrast coefficients tables

results$postHocs					an array of post-hoc tables

results$simpleEffects$Anova					a table of ANOVA for simple effects

results$simpleEffects$Params					a table

results$emeansTables					an array of predicted means tables

results$descPlot					a descriptives plot

results$descPlots					an array of results plots

results$plotnotes					a html

results$assumptions$normTest					a table of normality tests

results$assumptions$qq					a q-q plot

results$assumptions$normPlot					Residual histogram

results$assumptions$residPlot					Residual Predicted plot

results$assumptions$clusterBoxplot					Residuals boxplot by cluster

results$assumptions$randHist					an array of random coefficients histograms

Tables can be converted to data frames with asDF or as.data.frame. For example:

results$info$asDF

as.data.frame(results$info)

Examples

data(subjects_by_stimuli)

gamlj::gamljMixed(
       formula = y ~ 1 + cond+( 1|subj ),
       data = subjects_by_stimuli)
#> 
#>  MIXED MODEL
#> 
#>  Model Info                                                  
#>  ─────────────────────────────────────────────────────────── 
#>    Info                                                      
#>  ─────────────────────────────────────────────────────────── 
#>    Estimate                 Linear mixed model fit by REML   
#>    Call                     y ~ 1 + cond+( 1 | subj )        
#>    AIC                                      16561.77920000   
#>    BIC                                      16589.87570000   
#>    LogLikel.                                -8278.92510489   
#>    R-squared Marginal                           0.01264665   
#>    R-squared Conditional                        0.25385103   
#>    Converged                yes                              
#>    Optimizer                bobyqa                           
#>  ─────────────────────────────────────────────────────────── 
#> 
#> 
#>  MODEL RESULTS
#> 
#>  Fixed Effect Omnibus tests                               
#>  ──────────────────────────────────────────────────────── 
#>            F           Num df    Den df      p            
#>  ──────────────────────────────────────────────────────── 
#>    cond    50.83073         1    2949.000    < .0000001   
#>  ──────────────────────────────────────────────────────── 
#>    Note. Satterthwaite method for degrees of freedom
#> 
#> 
#>  Fixed Effects Parameter Estimates                                                                                
#>  ──────────────────────────────────────────────────────────────────────────────────────────────────────────────── 
#>    Names          Estimate      SE            Lower         Upper         df            t            p            
#>  ──────────────────────────────────────────────────────────────────────────────────────────────────────────────── 
#>    (Intercept)    19.6043644    0.30719153    19.0022801    20.2064487      49.00000    63.818051    < .0000001   
#>    cond            0.4849540    0.06802011     0.3516370     0.6182709    2949.00000     7.129567    < .0000001   
#>  ──────────────────────────────────────────────────────────────────────────────────────────────────────────────── 
#> 
#> 
#>  Random Components                                                 
#>  ───────────────────────────────────────────────────────────────── 
#>    Groups      Name           SD          Variance     ICC         
#>  ───────────────────────────────────────────────────────────────── 
#>    subj        (Intercept)    2.118253     4.486995    0.2442939   
#>    Residual                   3.725615    13.880206                
#>  ───────────────────────────────────────────────────────────────── 
#>    Note. Number of Obs: 3000 , groups: subj 50
#> 

gamlj::gamljMixed(
       data = subjects_by_stimuli,
       dep = "y",
       factors = "cond",
       modelTerms = "cond",
       cluster = "subj",
       randomTerms=list(list(c("cond","subj"))))
#> 
#>  MIXED MODEL
#> 
#>  Model Info                                                  
#>  ─────────────────────────────────────────────────────────── 
#>    Info                                                      
#>  ─────────────────────────────────────────────────────────── 
#>    Estimate                 Linear mixed model fit by REML   
#>    Call                     y ~ 1 + cond+( 0+cond | subj )   
#>    AIC                                           16415.340   
#>    BIC                                           16452.359   
#>    LogLikel.                                     -8202.161   
#>    R-squared Marginal                                  NaN   
#>    R-squared Conditional                               NaN   
#>    Converged                yes                              
#>    Optimizer                bobyqa                           
#>  ─────────────────────────────────────────────────────────── 
#>    Note. R-squared cannot be computed.
#> 
#> 
#>  MODEL RESULTS
#> 
#>  Fixed Effect Omnibus tests                              
#>  ─────────────────────────────────────────────────────── 
#>            F           Num df    Den df      p           
#>  ─────────────────────────────────────────────────────── 
#>    cond    9.032831         1    48.99999    0.0041723   
#>  ─────────────────────────────────────────────────────── 
#>    Note. Satterthwaite method for degrees of freedom
#> 
#> 
#>  Fixed Effects Parameter Estimates                                                                                           
#>  ─────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────── 
#>    Names          Effect         Estimate      SE           Lower         Upper        df          t            p            
#>  ─────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────── 
#>    (Intercept)    (Intercept)    19.6043644    0.3071916    19.0022800    20.206449    48.99999    63.818043    < .0000001   
#>    cond1          1 - -1          0.9699079    0.3227146     0.3373990     1.602417    48.99999     3.005467     0.0041723   
#>  ─────────────────────────────────────────────────────────────────────────────────────────────────────────────────────────── 
#> 
#> 
#>  Random Components                                      
#>  ────────────────────────────────────────────────────── 
#>    Groups      Name      SD          Variance     ICC   
#>  ────────────────────────────────────────────────────── 
#>    subj        cond-1    2.236543     5.002126          
#>                cond1     2.486998     6.185159          
#>    Residual              3.577006    12.794970          
#>  ────────────────────────────────────────────────────── 
#>    Note. Number of Obs: 3000 , groups: subj 50
#> 
#> 
#>  Random Parameters correlations                
#>  ───────────────────────────────────────────── 
#>    Groups    Param.1    Param.2    Corr.       
#>  ───────────────────────────────────────────── 
#>    subj      cond-1     cond1      0.6142310   
#>  ───────────────────────────────────────────── 
#>