GAMLj plot

This function re-estimates a GAMLj model adding a new plot. If no option is passed, extracts the plots present in the `gamlj*Results` object. If one plot is present, it is returned as a ggplot2 object, if more than one is present, a list of plots is returned. FALSE is returned if no plot is present or defined.

# S3 method for gamlj
plot(x, formula = NULL, ...)

Arguments

x

a gamlj results object of the class `gamlj`

formula

a right hand side formula specifying the effect to plot, of the form `~x`, `~x*z` or `~x*z*w`. It has prevalence on other options defining a plot.

...

all options accepted by a gamlj model function. Relevant for new plots are plot_x, plot_z and plot_by

Value

an object of class ggplot or a list of ggplot objects

Author

Marcello Gallucci

Examples

data(qsport)
gmod<-GAMLj3::gamlj_lm(
  formula = performance ~ hours,
  data = qsport)
#> 
#> MODULE: lm   #### phase init  ####
#> 
#> INIT TIME: 0.075732946395874  secs
#> 
#> MODULE:  #### phase run ####
#> 
#> MODULE: Estimating the model: checking
#> 
#> MODULE: Estimating the model: making options
#> 
#> MODULE: Estimating the model: running
#> 
#> RUNNER: initial estimation done,  0.00188207626342773  secs
#> 
#> GPARAMETERS: default .parameters for class lm
#> 
#> MODULE:  #### phase end ####
#> 
#> RUN TIME: 0.166595458984375  secs
#> 
#> TIME: 0.245238065719604  secs

plot(gmod,plot_x = 'hours')
#> 
#> MODULE: lm   #### phase init  ####
#> 
#> PLOTTER: init main plot
#> 
#> INIT TIME: 0.0744504928588867  secs
#> 
#> MODULE:  #### phase run ####
#> 
#> MODULE: Estimating the model: checking
#> 
#> MODULE: Estimating the model: making options
#> 
#> MODULE: Estimating the model: running
#> 
#> RUNNER: initial estimation done,  0.00194406509399414  secs
#> 
#> GPARAMETERS: default .parameters for class lm
#> 
#> PLOTTER: prepare main plot
#> 
#> MODULE:  #### phase end ####
#> 
#> RUN TIME: 0.201501369476318  secs
#> 
#> TIME: 0.278612613677979  secs
#> Error: 'descPlots' does not exist in this results element
plot(gmod,formula=~hours)
#> Error in (function (formula = NULL, data, dep = NULL, fixed_intercept = TRUE,     factors = NULL, covs = NULL, model_terms = NULL, nested_terms = NULL,     nested_intercept = NULL, omnibus = "F", estimates_ci = TRUE,     betas_ci = FALSE, ci_width = 95, ci_method = "wald", boot_r = 1000,     contrasts = NULL, show_contrastnames = TRUE, show_contrastcodes = FALSE,     vcov = FALSE, plot_x = NULL, plot_z = NULL, plot_by = NULL,     plot_raw = FALSE, plot_yscale = FALSE, plot_xoriginal = FALSE,     plot_black = FALSE, plot_around = "ci", emmeans = NULL, posthoc = NULL,     simple_x = NULL, simple_mods = NULL, simple_interactions = FALSE,     covs_scale = NULL, covs_conditioning = "mean_sd", ccm_value = 1,     ccp_value = 25, covs_scale_labels = "labels", adjust = list("bonf"),     posthoc_es = list("dm"), d_ci = FALSE, es = list("beta",         "etap"), homo_test = FALSE, qq_plot = FALSE, norm_test = FALSE,     norm_plot = FALSE, resid_plot = FALSE, intercept_info = FALSE,     es_info = FALSE, dep_scale = "none", se_method = "standard") {    if (!requireNamespace("jmvcore", quietly = TRUE))         stop("gamljGlm requires jmvcore to be installed (restart may be required)")    if (!missing(dep))         dep <- jmvcore::resolveQuo(jmvcore::enquo(dep))    if (!missing(factors))         factors <- jmvcore::resolveQuo(jmvcore::enquo(factors))    if (!missing(covs))         covs <- jmvcore::resolveQuo(jmvcore::enquo(covs))    if (!missing(plot_x))         plot_x <- jmvcore::resolveQuo(jmvcore::enquo(plot_x))    if (!missing(plot_z))         plot_z <- jmvcore::resolveQuo(jmvcore::enquo(plot_z))    if (!missing(plot_by))         plot_by <- jmvcore::resolveQuo(jmvcore::enquo(plot_by))    if (!missing(simple_x))         simple_x <- jmvcore::resolveQuo(jmvcore::enquo(simple_x))    if (!missing(simple_mods))         simple_mods <- jmvcore::resolveQuo(jmvcore::enquo(simple_mods))    if (missing(data))         data <- jmvcore::marshalData(parent.frame(), if (!missing(dep))             dep        else NULL, if (!missing(factors))             factors        else NULL, if (!missing(covs))             covs        else NULL, if (!missing(plot_x))             plot_x        else NULL, if (!missing(plot_z))             plot_z        else NULL, if (!missing(plot_by))             plot_by        else NULL, if (!missing(simple_x))             simple_x        else NULL, if (!missing(simple_mods))             simple_mods        else NULL)    for (v in factors) if (v %in% names(data))         data[[v]] <- as.factor(data[[v]])    .caller = "lm"    model_type = "lm"    .interface = "R"    donotrun = FALSE    if (inherits(model_terms, "formula")) {        f <- rFormula$new(model_terms, data)        model_terms <- f$terms        if (missing(fixed_intercept))             fixed_intercept <- f$intercept        if (missing(factors))             factors <- f$factors        if (missing(covs))             factors <- f$covs    }    if (!is.null(formula)) {        f <- rFormula$new(formula, data)        dep <- f$dep        factors <- f$factors        covs <- f$covs        fixed_intercept <- f$intercept        model_terms <- f$terms    }    if (is.null(model_terms))         model_terms <- as.list(c(factors, covs))    comparison = FALSE    if (inherits(nested_terms, "formula")) {        f <- rFormula$new(nested_terms)        nested_intercept <- f$intercept        nested_terms <- f$terms    }    if (is.something(nested_terms) | !is.null(nested_intercept))         comparison <- TRUE    if (inherits(emmeans, "formula"))         emmeans <- jmvcore::decomposeFormula(emmeans)    if (inherits(posthoc, "formula"))         posthoc <- jmvcore::decomposeFormula(posthoc)    if (is.something(names(covs_scale)))         covs_scale <- lapply(names(covs_scale), function(a) list(var = a,             type = covs_scale[[a]]))    if (is.something(names(contrasts)))         contrasts <- lapply(names(contrasts), function(a) list(var = a,             type = contrasts[[a]]))    if (se_method != "standard") {        se_method <- "robust"        robust_method <- se_method    }    options <- gamljlmOptions$new(.caller = .caller, .interface = .interface,         dep = dep, factors = factors, covs = covs, model_terms = model_terms,         nested_terms = nested_terms, comparison = comparison,         fixed_intercept = fixed_intercept, nested_intercept = nested_intercept,         omnibus = omnibus, estimates_ci = estimates_ci, betas_ci = betas_ci,         ci_width = ci_width, ci_method = ci_method, boot_r = boot_r,         contrasts = contrasts, show_contrastnames = show_contrastnames,         show_contrastcodes = show_contrastcodes, vcov = vcov,         plot_x = plot_x, plot_z = plot_z, plot_by = plot_by,         plot_raw = plot_raw, plot_yscale = plot_yscale, plot_xoriginal = plot_xoriginal,         plot_black = plot_black, plot_around = plot_around, emmeans = emmeans,         posthoc = posthoc, simple_x = simple_x, simple_mods = simple_mods,         simple_interactions = simple_interactions, covs_scale = covs_scale,         covs_conditioning = covs_conditioning, ccm_value = ccm_value,         ccp_value = ccp_value, covs_scale_labels = covs_scale_labels,         adjust = adjust, posthoc_es = posthoc_es, d_ci = d_ci,         model_type = model_type, es = es, homo_test = homo_test,         qq_plot = qq_plot, norm_test = norm_test, norm_plot = norm_plot,         resid_plot = resid_plot, intercept_info = intercept_info,         es_info = es_info, dep_scale = dep_scale, se_method = se_method)    analysis <- gamljlmClass$new(options = options, data = data)    analysis$run()    analysis$results})(dep = "performance", fixed_intercept = TRUE, factors = character(0),     covs = "hours", model_terms = list("hours"), omnibus = "F",     estimates_ci = TRUE, betas_ci = FALSE, ci_width = 95, ci_method = "wald",     boot_r = 1000, show_contrastnames = TRUE, show_contrastcodes = FALSE,     vcov = FALSE, plot_raw = FALSE, plot_yscale = FALSE, plot_xoriginal = FALSE,     plot_black = FALSE, plot_around = "ci", simple_interactions = FALSE,     covs_conditioning = "mean_sd", ccm_value = 1, ccp_value = 25,     covs_scale_labels = "labels", adjust = "bonf", posthoc_es = "dm",     d_ci = FALSE, es = c("beta", "etap"), homo_test = FALSE,     qq_plot = FALSE, norm_test = FALSE, norm_plot = FALSE, resid_plot = FALSE,     intercept_info = FALSE, es_info = FALSE, dep_scale = "none",     se_method = "standard", plotHAxis = "hours", data = structure(list(        performance = c(37, 32, 42, 35, 49, 38, 48, 38, 40, 41,         42, 22, 28, 43, 36, 43, 41, 33, 34, 47, 34, 33, 42, 43,         37, 43, 34, 30, 44, 23, 43, 46, 37, 33, 37, 37, 42, 42,         22, 42, 40, 39, 42, 40, 41, 36, 49, 40, 41, 15, 35, 34,         40, 44, 23, 36, 33, 47, 37, 26, 35, 39, 41, 42, 40, 37,         47, 30, 24, 47, 30, 35, 48, 38, 29, 31, 33, 40, 44, 37,         45, 32, 35, 40, 52, 47, 40, 37, 37, 43, 32, 21, 53, 43,         42, 41, 32, 46, 37, 35.001), hours = c(5, 5, 8, 8, 11,         8, 11, 5, 5, 11, 11, 2, 5, 11, 8, 14, 8, 5, 5, 8, 8,         8, 14, 8, 5, 11, 5, 5, 8, 2, 11, 8, 5, 5, 5, 8, 8, 8,         0, 11, 14, 11, 8, 8, 5, 11, 11, 8, 5, 2, 2, 8, 5, 5,         2, 14, 8, 8, 8, 14, 11, 8, 5, 8, 5, 8, 11, 5, 2, 11,         5, 5, 11, 11, 8, 8, 2, 8, 11, 5, 11, 5, 8, 8, 11, 11,         5, 8, 8, 11, 2, 2, 14, 8, 8, 8, 8, 11, 5, 11.001)), class = "data.frame", row.names = c(NA,     -100L))): unused argument (plotHAxis = "hours")