Post-hoc test on GAMLj results

This is a convenience function to re-estimates a GAMLj model adding posthoc tests. If no options is passed, extracts the post-hoc tests tables already in the model results (if any). If new post-hoc are defined, the post-hoc tests tables are returned.

posthoc(object, ...)

# S3 method for gamlj
posthoc(object, formula = NULL, ...)

Arguments

object

a gamlj results object of the class `gamlj`

...

all options accepted by a gamlj model function. Relevant for new tests are `postHoc` (a list of list of terms), `postHocCorr`, a list of correction to apply: one or more of 'none', 'bonf', or 'holm'; provide no, Bonferroni, and Holm Post Hoc corrections respectively.

formula

a right hand side formula specifying the factors or factors combinations to test, of the form `~x+z`, `~x:z` or `~x*z`. It has prevalence on other options defining a post-hoc test via character options.

Value

an object of class ggplot or a list of ggplot objects

Author

Marcello Gallucci

Examples

data(fivegroups)
fivegroups$Group<-factor(fivegroups$Group)
gmod<-gamlj::gamljGlm(
  formula = Score ~Group,
  data = fivegroups)

posthoc(gmod,formula =~Group)
#> Error in (function (data, formula, dep = NULL, factors = NULL, covs = NULL,     modelTerms = NULL, fixedIntercept = TRUE, showParamsCI = TRUE,     paramCIWidth = 95, contrasts = NULL, showRealNames = TRUE,     showContrastCode = FALSE, plotHAxis = NULL, plotSepLines = NULL,     plotSepPlots = NULL, plotRaw = FALSE, plotDvScale = FALSE,     plotError = "none", ciWidth = 95, postHoc = NULL, eDesc = FALSE,     eCovs = FALSE, simpleVariable = NULL, simpleModerator = NULL,     simple3way = NULL, simpleScale = "mean_sd", cvalue = 1, percvalue = 25,     simpleScaleLabels = "labels", postHocCorr = list("bonf"),     scaling = NULL, effectSize = list("beta", "etap"), homoTest = FALSE,     qq = FALSE, normTest = FALSE, normPlot = FALSE, residPlot = FALSE,     interceptInfo = FALSE, effectSizeInfo = FALSE, dep_scale = "none") {    if (!requireNamespace("jmvcore"))         stop("gamljGlm requires jmvcore to be installed (restart may be required)")    if (!missing(formula)) {        if (missing(dep))             dep <- jmvcore::marshalFormula(formula = formula,                 data = if (!missing(data))                   data                else NULL, from = "lhs", subset = "1", required = TRUE)        if (missing(factors))             factors <- jmvcore::marshalFormula(formula = formula,                 data = if (!missing(data))                   data                else NULL, from = "rhs", type = "vars", permitted = "factor")        if (missing(covs))             covs <- jmvcore::marshalFormula(formula = formula,                 data = if (!missing(data))                   data                else NULL, from = "rhs", type = "vars", permitted = "numeric")        if (missing(modelTerms))             modelTerms <- jmvcore::marshalFormula(formula = formula,                 data = if (!missing(data))                   data                else NULL, from = "rhs", type = "terms")    }    if (!missing(dep))         dep <- jmvcore::resolveQuo(jmvcore::enquo(dep))    if (!missing(factors))         factors <- jmvcore::resolveQuo(jmvcore::enquo(factors))    if (!missing(covs))         covs <- jmvcore::resolveQuo(jmvcore::enquo(covs))    if (!missing(plotHAxis))         plotHAxis <- jmvcore::resolveQuo(jmvcore::enquo(plotHAxis))    if (!missing(plotSepLines))         plotSepLines <- jmvcore::resolveQuo(jmvcore::enquo(plotSepLines))    if (!missing(plotSepPlots))         plotSepPlots <- jmvcore::resolveQuo(jmvcore::enquo(plotSepPlots))    if (!missing(simpleVariable))         simpleVariable <- jmvcore::resolveQuo(jmvcore::enquo(simpleVariable))    if (!missing(simpleModerator))         simpleModerator <- jmvcore::resolveQuo(jmvcore::enquo(simpleModerator))    if (!missing(simple3way))         simple3way <- jmvcore::resolveQuo(jmvcore::enquo(simple3way))    if (missing(data))         data <- jmvcore::marshalData(parent.frame(), if (!missing(dep))             dep        else NULL, if (!missing(factors))             factors        else NULL, if (!missing(covs))             covs        else NULL, if (!missing(plotHAxis))             plotHAxis        else NULL, if (!missing(plotSepLines))             plotSepLines        else NULL, if (!missing(plotSepPlots))             plotSepPlots        else NULL, if (!missing(simpleVariable))             simpleVariable        else NULL, if (!missing(simpleModerator))             simpleModerator        else NULL, if (!missing(simple3way))             simple3way        else NULL)    for (v in factors) if (v %in% names(data))         data[[v]] <- as.factor(data[[v]])    if (inherits(modelTerms, "formula"))         modelTerms <- jmvcore::decomposeFormula(modelTerms)    if (inherits(postHoc, "formula"))         postHoc <- jmvcore::decomposeFormula(postHoc)    if (is.something(names(scaling)))         scaling <- lapply(names(scaling), function(a) list(var = a,             type = scaling[[a]]))    if (is.something(names(contrasts)))         contrasts <- lapply(names(contrasts), function(a) list(var = a,             type = contrasts[[a]]))    options <- gamljGlmOptions$new(dep = dep, factors = factors,         covs = covs, modelTerms = modelTerms, fixedIntercept = fixedIntercept,         showParamsCI = showParamsCI, paramCIWidth = paramCIWidth,         contrasts = contrasts, showRealNames = showRealNames,         showContrastCode = showContrastCode, plotHAxis = plotHAxis,         plotSepLines = plotSepLines, plotSepPlots = plotSepPlots,         plotRaw = plotRaw, plotDvScale = plotDvScale, plotError = plotError,         ciWidth = ciWidth, postHoc = postHoc, eDesc = eDesc,         eCovs = eCovs, simpleVariable = simpleVariable, simpleModerator = simpleModerator,         simple3way = simple3way, simpleScale = simpleScale, cvalue = cvalue,         percvalue = percvalue, simpleScaleLabels = simpleScaleLabels,         postHocCorr = postHocCorr, scaling = scaling, effectSize = effectSize,         homoTest = homoTest, qq = qq, normTest = normTest, normPlot = normPlot,         residPlot = residPlot, interceptInfo = interceptInfo,         effectSizeInfo = effectSizeInfo, dep_scale = dep_scale)    analysis <- gamljGlmClass$new(options = options, data = data)    analysis$run()    analysis$results})(dep = "Score", factors = "Group", covs = character(0), modelTerms = list(    "Group"), fixedIntercept = TRUE, showParamsCI = TRUE, paramCIWidth = 95,     showRealNames = TRUE, showContrastCode = FALSE, plotRaw = FALSE,     plotDvScale = FALSE, plotError = "none", ciWidth = 95, eDesc = FALSE,     eCovs = FALSE, simpleScale = "mean_sd", cvalue = 1, percvalue = 25,     simpleScaleLabels = "labels", postHocCorr = "bonf", effectSize = c("beta",     "etap"), homoTest = FALSE, qq = FALSE, normTest = FALSE,     normPlot = FALSE, residPlot = FALSE, interceptInfo = FALSE,     effectSizeInfo = FALSE, dep_scale = "none", posthoc = ~Group,     data = structure(list(Score = c(0.12, 0.8, -0.27, 1.04, -0.19,     0.5, -0.62, 1.2, -0.46, 0.77, -0.08, 0.23, 0.19, 0.03, 0.04,     -0.04, 0.85, -0.27, -0.23, 0.35, -0.27, -0.19, 0.07, 1.04,     0.35, 0.19, -0.27, 1.46, -0.15, -0.65, 0.15, 0.19, 0.38,     -0.35, 0.77, 0.19, -0.19, -1.81, -0.5, 0.16, 0.08, 0.34,     -0.12, -0.61, -0.65, 0, 0.23, -0.15, 0.39, 1.38, 0.81, 0.31,     0.34, 0.61, 0.19, 0.54, 0.08, 0.58, 0.5, 0.85, -0.3, -0.15,     0.12, 0.38, 0.46, 1.15, 0.16, 0.24, 0.38, 0.7, -0.85, -0.12,     0, -0.39, 1.27, -0.5, 0.73, 0.08, 0.08, -0.35, -0.69, 0.23,     0.11, -0.27, -0.03, 0.62, -0.07, 0.07, -0.53, 0.04, 0.19,     -0.84, 0.57, -0.16, -0.3, -0.04, 1, 0.08, 0.04, 2.11, 0.5,     0.12, 0.12, -0.27, 0.61, 0.39, 0.5, 0.96, 1.54, -0.81, 0.77,     -0.04, 0.46, -0.2, -0.42, 0.81, -0.08, 0.88, -0.08, 0.27,     0.57, 0, 0.39, -0.23, -0.34, 0.23, 0, 0.5, 1, -0.81, -0.23,     0.15, 0.85, 1.23, 0.27, -0.5, -0.38, -0.15, 0.27, -0.23,     0.24, 1.62, -0.46, -0.23, 0.58, -0.15, 0.58, -0.27, 0.34,     1.08, -0.07, -0.08, -0.27, 0.5, 0.15, -0.31, -0.35, 0.27,     -0.15, 0.43, -0.35, 0.54, 0.15, 0.74, -0.19, 0.16, -0.96,     0.35, 0.39, 0.57), Group = structure(c(3L, 4L, 1L, 3L, 4L,     1L, 2L, 3L, 1L, 4L, 4L, 3L, 2L, 3L, 1L, 3L, 4L, 4L, 2L, 1L,     2L, 2L, 3L, 4L, 4L, 4L, 1L, 4L, 1L, 1L, 1L, 1L, 3L, 1L, 3L,     3L, 3L, 3L, 1L, 1L, 2L, 3L, 4L, 2L, 2L, 1L, 2L, 1L, 3L, 3L,     2L, 3L, 1L, 3L, 2L, 3L, 1L, 3L, 4L, 3L, 2L, 1L, 4L, 4L, 4L,     3L, 1L, 2L, 2L, 4L, 2L, 1L, 3L, 3L, 3L, 4L, 4L, 3L, 1L, 4L,     3L, 1L, 4L, 1L, 2L, 1L, 1L, 2L, 3L, 2L, 1L, 1L, 4L, 1L, 1L,     2L, 4L, 2L, 4L, 4L, 4L, 1L, 2L, 2L, 4L, 1L, 3L, 3L, 4L, 4L,     1L, 1L, 3L, 2L, 2L, 4L, 2L, 4L, 2L, 4L, 4L, 3L, 3L, 4L, 1L,     3L, 3L, 4L, 4L, 2L, 3L, 3L, 4L, 1L, 3L, 4L, 2L, 2L, 4L, 1L,     3L, 4L, 3L, 4L, 4L, 2L, 3L, 1L, 2L, 4L, 1L, 3L, 1L, 2L, 4L,     1L, 3L, 3L, 4L, 3L, 2L, 2L, 3L, 1L, 1L, 1L, 2L, 1L, 4L, 2L    ), .Label = c("1", "2", "3", "4"), class = "factor")), class = "data.frame", row.names = c(NA,     -170L))): unused argument (posthoc = ~Group)