S3 methods for class jamovi ResultsElement

These functions extract all visible tables from a ResultsElement or related classes produced by GAMLj3 and print them in R style.

These functions extract all visible tables from a list of tables produced by GAMLj3 and print them in R style.

# S3 method for class 'ResultsElement'
summary(object, ...)

# S3 method for class 'gamlj_list'
summary(object, ...)

# S3 method for class 'jmvrtable'
print(x, ...)

# S3 method for class 'jmvrobj'
print(x, ...)

# S3 method for class 'gamlj_list'
print(x, ...)

# S3 method for class 'gamlj'
coef(object, ...)

fit(x, ...)

# S3 method for class 'gamlj'
fit(x, ...)

# S3 method for class 'gamlj'
em_means(object, formula = NULL, ...)

Arguments

object

a gamlj results object of the class `gamlj`

...

not used

x

a gamlj results object of the class `gamlj`

formula

not used

Value

a list of table as data.frame

a list of tables as data.frame

Author

Marcello Gallucci

Examples

data(fivegroups)
fivegroups$Group<-factor(fivegroups$Group)
gmod<-GAMLj3::gamlj_lm(
  formula = Score ~Group,
  data = fivegroups)

summary(gmod)
#> Model Info 
#> 
#>            info        value                            specs
#> 1    Model Type Linear Model       OLS Model for continuous y
#> 2         Model           lm                Score ~ 1 + Group
#> 3  Distribution     Gaussian Normal distribution of residuals
#> 4 Omnibus Tests            F                                 
#> 5   Sample size          170                                 
#> 6     Converged          yes                                 
#> 7   Y transform         none                                 
#> 8   C.I. method         Wald                                 
#> 
#> 
#> Model Fit 
#> 
#>          r2       ar2 df1 df2        f            p
#> 1 0.1472811 0.1318705   3 166 9.557137 7.399073e-06
#> 
#> 
#> ANOVA Omnibus tests 
#> 
#>      source        ss  df        f            p    etaSqP
#> 1     Model  7.483173   3 9.557137 7.399073e-06 0.1472811
#> 2     Group  7.483173   3 9.557137 7.399073e-06 0.1472811
#> 3 Residuals 43.325624 166       NA           NA        NA
#> 4     Total 50.808798 169       NA           NA        NA
#> 
#> 
#> Parameter Estimates (Coefficients) 
#> 
#>        source       label   estimate         se est.ci.lower est.ci.upper
#> 1 (Intercept) (Intercept)  0.1561919 0.03935543   0.07849021    0.2338936
#> 2      Group1       2 - 1 -0.1001010 0.11481160  -0.32678019    0.1265782
#> 3      Group2       3 - 1  0.2435657 0.10831305   0.02971695    0.4574144
#> 4      Group3       4 - 1  0.4431212 0.10831305   0.22927250    0.6569699
#>          beta  df       test            p
#> 1 -0.02539755 166  3.9687510 1.073436e-04
#> 2 -0.18256283 166 -0.8718719 3.845375e-01
#> 3  0.44421165 166  2.2487194 2.584581e-02
#> 4  0.80815829 166  4.0911157 6.682737e-05
#> 
#> 
data(fivegroups)
fivegroups$Group<-factor(fivegroups$Group)
gmod<-GAMLj3::gamlj_lm(
  formula = Score ~Group,
  data = fivegroups)

summary(gmod)
#> Model Info 
#> 
#>            info        value                            specs
#> 1    Model Type Linear Model       OLS Model for continuous y
#> 2         Model           lm                Score ~ 1 + Group
#> 3  Distribution     Gaussian Normal distribution of residuals
#> 4 Omnibus Tests            F                                 
#> 5   Sample size          170                                 
#> 6     Converged          yes                                 
#> 7   Y transform         none                                 
#> 8   C.I. method         Wald                                 
#> 
#> 
#> Model Fit 
#> 
#>          r2       ar2 df1 df2        f            p
#> 1 0.1472811 0.1318705   3 166 9.557137 7.399073e-06
#> 
#> 
#> ANOVA Omnibus tests 
#> 
#>      source        ss  df        f            p    etaSqP
#> 1     Model  7.483173   3 9.557137 7.399073e-06 0.1472811
#> 2     Group  7.483173   3 9.557137 7.399073e-06 0.1472811
#> 3 Residuals 43.325624 166       NA           NA        NA
#> 4     Total 50.808798 169       NA           NA        NA
#> 
#> 
#> Parameter Estimates (Coefficients) 
#> 
#>        source       label   estimate         se est.ci.lower est.ci.upper
#> 1 (Intercept) (Intercept)  0.1561919 0.03935543   0.07849021    0.2338936
#> 2      Group1       2 - 1 -0.1001010 0.11481160  -0.32678019    0.1265782
#> 3      Group2       3 - 1  0.2435657 0.10831305   0.02971695    0.4574144
#> 4      Group3       4 - 1  0.4431212 0.10831305   0.22927250    0.6569699
#>          beta  df       test            p
#> 1 -0.02539755 166  3.9687510 1.073436e-04
#> 2 -0.18256283 166 -0.8718719 3.845375e-01
#> 3  0.44421165 166  2.2487194 2.584581e-02
#> 4  0.80815829 166  4.0911157 6.682737e-05
#> 
#>