Simple Effects on GAMLj results

This is a convenience function to re-estimates a GAMLj model adding simple effect analysis. If no options is passed, extracts the simple effects tables already in the model results (if any). If new tests are defined, the simple effects tests tables are returned.

simpleEffects(object, ...)

# S3 method for gamlj
simpleEffects(object, formula = NULL, ...)

Arguments

object

a gamlj results object of the class `gamlj`

...

all options accepted by a gamlj model function. Relevant for new tests are `simpleVariable` (the simple effect variable), `simpleModerator` (the moderator), and `simple3way` for the second moderator.

formula

a right hand side formula specifying the variables to test, of the form `~x:z`, `~x:z:w` or `~x*z`. The formula is not exanded, so the first variable is the simple effect variable, the second is the moderator, the third an optional additional moderator It has prevalence on other options defining a simple effects test via character options.

Value

an object of class gamlj results

Author

Marcello Gallucci

Examples

data(winkel)
#> Warning: data set ‘winkel’ not found
wicksell$time<-factor(wicksell$time)
wicksell$group<-factor(wicksell$group) 
gmod<-gamlj::gamljMixed(
   formula = dv ~ 1 +group+ time:group+ time+( 1 | subj ),
   data = wicksell)

simpleEffects(gmod,formula =~time:group)
#> Error in simpleEffects(gmod, formula = ~time:group): could not find function "simpleEffects"